Fix Eslint errors
