+++ index.css
@@ -142,7 +142,6 @@ form h2 {
   justify-content: flex-end;
   margin-right: 20px;
   font-size: 16px;
-  /* position: fixed; */
 }
 
 .cont {
