+++ modules/bookCollection.js
@@ -1,0 +1,29 @@
+let bookCollection = JSON.parse(localStorage.getItem('bookCollection')) || [];
+
+const removeBook = (book) => {
+  bookCollection = bookCollection.filter((b) => b !== book);
+  localStorage.setItem('bookCollection', JSON.stringify(bookCollection));
+};
+const displayBooks = (container) => {
+  container.innerHTML = '';
+  bookCollection.forEach((book) => {
+    const newBook = document.createElement('div');
+    newBook.classList.add('newBook');
+    newBook.innerHTML = `<p>${book.title} by ${book.author}</p><button class="remove">Remove</button>`;
+    container.appendChild(newBook);
+
+    const removeBtn = newBook.querySelector('.remove');
+    removeBtn.addEventListener('click', () => {
+      removeBook(book);
+      newBook.remove();
+    });
+  });
+};
+
+const addBook = (title, author) => {
+  const newBook = { title, author };
+  bookCollection.push(newBook);
+  localStorage.setItem('bookCollection', JSON.stringify(bookCollection));
+};
+
+export { displayBooks, addBook };
