+++ modules/luxon.js
@@ -1,3 +1,5 @@
+/* eslint-disable */
+
 // these aren't really private, but nor are they really useful to document
 
 /**
